Shooting and Input Tweaks
Added shooting
Players can shoot and record the shots (only record successful shots)
Follower can read the shoot memory to shoot later on.

I want to use the PlayerController class to control the player and the
shadow. So I've started to do some of the tweaks to make that work.
PlayerInput class handles the input instead so the controller script can
be more abstract. It should probably have a different name someday

The action memory in the PlayerController works a little differently than
the movement mem. It only creates new items on successful actions instead
of one every frame. You don't shoot every frame, but you do move every
frame.

Follower
Made it work with the PlayerController.
It controls the logic and when to do things. Like a brain or input from
the player.
Instead of checking each frame if it has an action from the action memory
to enact, it just queues the next one up with a coroutine.

New Stuff
see-through mat for the shost
tried to use scriptable objects, but it didn't work
guns and bullets
PlayerInput
ActionInTime - memory object for shooting instead of movement

Bugs
when the player holds down the mouse to shoot multiple shots in a row,
they do get saved correctly, but the shadow won't shoot all of the shots.
Might be just a slight timing issue where the gun's fire rate hasn't quite
finished. It's fire rate is 0.25, but I'm trying to shoot every
0.2499999s.